refactor(web): move error service to @safe-global/utils
errors
