+++ apps/web/src/components/settings/PushNotifications/hooks/useNotificationRegistrations.ts
@@ -8,7 +8,7 @@ import { trackEvent } from '@/services/analytics'
 import { PUSH_NOTIFICATION_EVENTS } from '@/services/analytics/events/push-notifications'
 import { getRegisterDevicePayload } from '../logic'
 import { logError } from '@/services/exceptions'
-import ErrorCodes from '@/services/exceptions/ErrorCodes'
+import ErrorCodes from '@safe-global/utils/services/exceptions/ErrorCodes'
 import useWallet from '@/hooks/wallets/useWallet'
 import type { NotifiableSafes } from '../logic'
 import { NotificationsTokenVersion } from '@/services/push-notifications/preferences'
