+++ apps/web/src/components/tx-flow/flows/SafeAppsTx/ConfirmSafeAppsTxDetails.tsx
@@ -6,7 +6,7 @@ import useWallet from '@/hooks/wallets/useWallet'
 import useOnboard from '@/hooks/wallets/useOnboard'
 import { dispatchSafeAppsTx } from '@/services/tx/tx-sender'
 import { trackSafeAppTxCount } from '@/services/safe-apps/track-app-usage-count'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import type { SafeAppsTxParams } from '.'
 
 export const ConfirmSafeAppsTxDetails = ({
