+++ apps/web/src/components/tx/SignOrExecuteForm/ExecuteThroughRoleForm/index.tsx
@@ -14,7 +14,7 @@ import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import { TxModalContext } from '@/components/tx-flow'
 import { SuccessScreenFlow } from '@/components/tx-flow/flows'
 import AdvancedParams, { useAdvancedParams } from '../../AdvancedParams'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import { isWalletRejection } from '@/utils/wallets'
 
 import css from '../styles.module.css'
