+++ apps/web/src/features/multichain/hooks/useSafeCreationData.ts
@@ -8,8 +8,8 @@ import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import { useAppSelector } from '@/store'
 import { determineMasterCopyVersion, isPredictedSafeProps } from '@/features/counterfactual/utils'
 import { logError } from '@/services/exceptions'
-import ErrorCodes from '@/services/exceptions/ErrorCodes'
-import { asError } from '@/services/exceptions/utils'
+import ErrorCodes from '@safe-global/utils/services/exceptions/ErrorCodes'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import semverSatisfies from 'semver/functions/satisfies'
 import { ZERO_ADDRESS } from '@safe-global/protocol-kit/dist/src/utils/constants'
 import { getSafeToL2SetupDeployment } from '@safe-global/safe-deployments'
