+++ apps/web/src/features/spaces/components/SignInButton/index.tsx
@@ -6,7 +6,7 @@ import { useAppDispatch } from '@/store'
 import { setAuthenticated } from '@/store/authSlice'
 import { showNotification } from '@/store/notificationsSlice'
 import { logError } from '@/services/exceptions'
-import ErrorCodes from '@/services/exceptions/ErrorCodes'
+import ErrorCodes from '@safe-global/utils/services/exceptions/ErrorCodes'
 
 const SignInButton = () => {
   const dispatch = useAppDispatch()
