+++ apps/web/src/features/speedup/components/SpeedUpModal.tsx
@@ -12,7 +12,7 @@ import { createExistingTx, dispatchCustomTxSpeedUp, dispatchSafeTxSpeedUp } from
 import { showNotification } from '@/store/notificationsSlice'
 import { useCallback, useState } from 'react'
 import GasParams from '@/components/tx/GasParams'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import { getTxOptions } from '@/utils/transactions'
 import { useCurrentChain, useHasFeature } from '@/hooks/useChains'
 import { SimpleTxWatcher } from '@/utils/SimpleTxWatcher'
@@ -25,7 +25,7 @@ import { MODALS_EVENTS, trackEvent } from '@/services/analytics'
 import { TX_EVENTS } from '@/services/analytics/events/transactions'
 import { getTransactionTrackingType } from '@/services/analytics/tx-tracking'
 import { trackError } from '@/services/exceptions'
-import ErrorCodes from '@/services/exceptions/ErrorCodes'
+import ErrorCodes from '@safe-global/utils/services/exceptions/ErrorCodes'
 import CheckWallet from '@/components/common/CheckWallet'
 import { useLazyGetTransactionDetailsQuery } from '@/store/api/gateway'
 import NetworkWarning from '@/components/new-safe/create/NetworkWarning'
