+++ apps/web/src/features/walletconnect/components/WcInput/index.tsx
@@ -5,7 +5,7 @@ import { WalletConnectContext } from '@/features/walletconnect/WalletConnectCont
 import useDebounce from '@/hooks/useDebounce'
 import { trackEvent } from '@/services/analytics'
 import { WALLETCONNECT_EVENTS } from '@/services/analytics/events/walletconnect'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import { getClipboard, isClipboardSupported } from '@/utils/clipboard'
 import { Button, CircularProgress, InputAdornment, TextField } from '@mui/material'
 import { useCallback, useContext, useEffect, useState } from 'react'
