+++ apps/web/src/hooks/safe-apps/useSafeAppFromManifest.ts
@@ -5,7 +5,7 @@ import { fetchSafeAppFromManifest } from '@/services/safe-apps/manifest'
 import useAsync from '@/hooks/useAsync'
 import { getEmptySafeApp } from '@/components/safe-apps/utils'
 import type { SafeAppDataWithPermissions } from '@/components/safe-apps/types'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 
 type UseSafeAppFromManifestReturnType = {
   safeApp: SafeAppDataWithPermissions
