+++ apps/web/src/services/tx/tx-sender/dispatch.ts
@@ -33,7 +33,7 @@ import {
   prepareApproveTxHash,
 } from './sdk'
 import { createWeb3, getUserNonce } from '@/hooks/wallets/web3'
-import { asError } from '@/services/exceptions/utils'
+import { asError } from '@safe-global/utils/services/exceptions/utils'
 import chains from '@/config/chains'
 import { createExistingTx } from './create'
