+++ packages/utils/src/services/exceptions/utils.ts
